https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved inboxCustomRoleMailboxService to esn.inbox.libs
