+++ src/linagora.esn.unifiedinbox/app/services/jmap-client-provider/jmap-client-provider.js
@@ -1,7 +1,7 @@
 require('../config/config.js');
 require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../../../esn.inbox.libs/app/services/generate-jwt-token/generate-jwt-token.js');
-require('../custom-role-mailbox/custom-role-mailbox.service.js');
+require('../../../../esn.inbox.libs/app/services/custom-role-mailbox/custom-role-mailbox.service.js');
 
 (function(angular) {
   'use strict';
